requirements.txt
aiosqlite==0.17.0
click==7.1.2
fastapi==0.65.1
h11==0.12.0
httptools==0.1.2
pydantic==1.8.2
python-dotenv==0.17.1
PyYAML==5.4.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
python-multipart==0.0.5
six==1.16.0
py-bcrypt==0.4